move TranscriptWord to its own file
