+++ reascripts/ReaSpeech/tests/TestTranscript.lua
@@ -9,6 +9,7 @@ require('mock_reaper')
 require('Polo')
 require('source/Transcript')
 require('source/TranscriptSegment')
+require('source/TranscriptWord')
 
 --
